refact(h2): move files and simplify h2 config
