+++ src/main/java/dev/dluks/contactsmanager/db/IDatabase.java
@@ -1,4 +1,4 @@
-package dev.dluks.contactsmanager;
+package dev.dluks.contactsmanager.db;
 
 import java.sql.SQLException;
