--- src/test/resources/data.sql
@@ -1,2 +1,0 @@
-INSERT INTO users (name, email, password) VALUES ('Alice', 'alice@example.com', 'hashedpassword1');
-INSERT INTO users (name, email, password) VALUES ('Bob', 'bob@example.com', 'hashedpassword2');
